changed netbeans configuration: added new files, tests changing type to 64bit in the Timer.wqhh
git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/branches/hash-v2@8831 a23ab9f9-1ea2-4934-960d-990259f9ea8c

+++ src/karabo/util/Timer.hh
@@ -39,8 +39,8 @@ namespace karabo {
             uint64_t epoch;
 
             struct {
-                uint32_t nsec;
-                uint32_t sec;
+                uint64_t nsec;
+                uint64_t sec;
             };
         } timestamp;
